refactor: 자동차 검증 로직을 domain/Car 와 domain/Cars 로 이동
