+++ src/main/java/racingcar/domain/Car.java
@@ -1,23 +1,46 @@
 package racingcar.domain;
 
 public class Car implements Comparable<Car> {
+	private static final int CAR_LENGTH_LIMIT = 5;
 	private static final String STEP = "-";
 
 	private final String name;
 	private int position = 0;
 
 	public Car(String name) {
+		checkValid(name);
 		this.name = name;
-	}
-
-	private void move() {
-		position++;
 	}
 
 	public void drive(boolean directing) {
 		if (directing) {
 			move();
 		}
+	}
+
+	private void checkValid(String names) {
+		checkBlank(names);
+		checkNameLength(names);
+	}
+
+	private void checkBlank(String name) {
+		String text = "자동차 이름은 %s일 수 없습니다.";
+		if (name == null) {
+			throw new NullPointerException(String.format(text, "null"));
+		}
+		if (name.trim().equals("")) {
+			throw new IllegalArgumentException(String.format(text, "공백"));
+		}
+	}
+
+	private void checkNameLength(String name) {
+		if (!(name.trim().length() <= CAR_LENGTH_LIMIT)) {
+			throw new IllegalArgumentException("자동차의 이름은 " + CAR_LENGTH_LIMIT + "글자를 초과할 수 없습니다.");
+		}
+	}
+
+	private void move() {
+		position++;
 	}
 
 	public boolean isSamePosition(Car other) {
